version 2/requirements.txt
streamlit==1.28.1
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
requests==2.31.0
python-dotenv==1.0.0